+++ CHANGELOG.md
@@ -3,10 +3,17 @@
 #### Bugfixes
 
 - Dungeon portals logic simplified and corrected
--
+- Fixed an issue with applying frame customizations on the Mythic+ frame not causing them to take effect
+- Fixed several issues with Mythic+ run logging
+- Fixed an issue where when opening the Mythic+ frame the first time, the Summary tab is not highlighted as active
 
 #### Improvements/New Features
 
 - BiS will now display notifications when entering dungeons or raids that drop BiS items
 - BiS will now display a notification if a BiS item is dropped and being rolled on
 - BiS will now display a notification (and highlight the item) if you open the Great Vault and a BiS item is available
+- Improved the Run Details frame that appears when you select a run
+- Clicking the logo in the top left of the Mythic+ frame will open the addon configuration
+- Added "Other", "Crafted", and "World Boss" as options to custom item sources in the BiS frame
+- Added a new "overall statistics" function to the Mythic+ addon, showing data such as success rate, runs, total time, deaths, group comps, etc.
+  - Added a new section to the Dungeons info panel that includes some of this data; a pie chart showing success rate, total runs, total time, total deaths. Includes a more info button that shows group comp breakdown, loot.